+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## v2.0.19.2
 - Avoid use enabled property without checking the existence
+- Monitor ad errors in honeybadger tool
 
 ## v2.0.19.1
 - Fix importing share function depending on presence of any cuepoints